Dynamic DCoffset removal for current. EmonLibCM_Vrms as double.
DCoffset for current needs to be eliminated the same way as for voltage.
Sequencing in phaseCal calculation corrected.
Cast added to avoid warnings. sampleV_minusDC_long moved to global.
sampleV_minusDC_long inited in EmonLibCM_Init()

+++ EmonLibCM.h
@@ -55,6 +55,6 @@ extern volatile boolean datalogEventPending;
 
 extern bool EmonLibCM_ACAC;
 
-extern int EmonLibCM_Vrms;
+extern double EmonLibCM_Vrms;
 
 #endif
